Lots of changes and more to come

+++ src/main/java/uk/ac/cam/sup/controllers/SubmissionController.java
@@ -4,11 +4,14 @@ import com.google.common.collect.ImmutableMap;
 import org.hibernate.Session;
 import org.jboss.resteasy.annotations.providers.multipart.MultipartForm;
 import uk.ac.cam.sup.HibernateUtil;
+import uk.ac.cam.sup.exceptions.MetadataNotFoundException;
 import uk.ac.cam.sup.forms.FileUploadForm;
 import uk.ac.cam.sup.helpers.UserHelper;
+import uk.ac.cam.sup.models.Answer;
 import uk.ac.cam.sup.models.Bin;
 import uk.ac.cam.sup.models.Submission;
 import uk.ac.cam.sup.tools.FilesManip;
+import uk.ac.cam.sup.tools.PDFManip;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -17,14 +20,11 @@ import java.io.File;
 import java.io.IOException;
 import java.util.*;
 
-import static uk.ac.cam.sup.tools.PDFManip.PdfAddHeader;
-import static uk.ac.cam.sup.tools.PDFManip.PdfMetadataInject;
-
 @Path ("/submission/{binId}")
+@Produces("application/json")
 public class SubmissionController {
 
     @POST
-    @Produces("application/json")
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     public Object createSubmission(@MultipartForm FileUploadForm uploadForm, @PathParam("binId") long binId) {
 
@@ -51,7 +51,7 @@ public class SubmissionController {
         session = HibernateUtil.getSession();
         session.beginTransaction();
 
-        String directory = "temp/" + user + "/submissions/answer/";
+        String directory = "temp/" + user + "/submissions/answers/";
         String fileName = "submission_" + submission.getId() + ".pdf";
 
         try {
@@ -62,22 +62,25 @@ public class SubmissionController {
             return Response.status(500).build();
         }
 
-        // todo: convert the received file;
-
-        // todo: split the received file;
-
-        // Fixme: Proper injections
-        PdfMetadataInject("users", "1", directory + fileName);
-        PdfMetadataInject("user.1", user, directory + fileName);
-        PdfMetadataInject("bin", Long.toString(binId), directory + fileName);
-
-        PdfAddHeader(directory + fileName, directory + "Headed" + fileName);
-
         submission.setBin(bin);
         submission.setOwner(user);
         submission.setFilePath(directory + fileName);
 
         session.update(submission);
+
+        // todo: convert the received file;
+
+        PDFManip.PdfMetadataInject("uploader", user, directory + fileName);
+
+        // ToDo: Redirect to splitting screen
+
+        PDFManip.PdfMetadataInject("page.1", "qqq 1", directory + fileName);
+        PDFManip.PdfMetadataInject("page.2", "qqq 1", directory + fileName);
+        PDFManip.PdfMetadataInject("page.3", "ExampleSheet", directory + fileName);
+        PDFManip.PdfMetadataInject("page.4", "qqq 5", directory + fileName);
+        PDFManip.PdfMetadataInject("page.5", "qqq 5", directory + fileName);
+
+        FilesManip.distributeSubmission(submission);
 
         return ImmutableMap.of("submission", ImmutableMap.of("id", submission.getId(),
                                                              "link", submission.getLink(),
@@ -85,7 +88,6 @@ public class SubmissionController {
     }
 
     @GET
-    @Produces("application/json")
     public Object listSubmissions(@PathParam("binId") long binId) {
 
         // Set Hibernate and get user
@@ -125,7 +127,6 @@ public class SubmissionController {
 
     @GET
     @Path("/{submissionId}")
-    @Produces("application/pdf")
     public Object seeSubmission(@PathParam("binId") long binId, @PathParam("submissionId") long submissionId) {
 
         // Set Hibernate and get user
@@ -151,7 +152,6 @@ public class SubmissionController {
 
     @DELETE
     @Path("/{submissionId}")
-    @Produces("application/json")
     public Object deleteSubmission(@PathParam("binId") long binId, @PathParam("submissionId") long submissionId) {
 
         // Set Hibernate and get user
@@ -167,10 +167,13 @@ public class SubmissionController {
 
         Submission submission = (Submission) session.get(Submission.class, submissionId);
 
-        if (!bin.canDeleteSubmission(user, submission))
-            return Response.status(401).build();
+        if (session != null)
+        {
+            if (!bin.canDeleteSubmission(user, submission))
+                return Response.status(401).build();
 
-        session.delete(submission);
+            session.delete(submission);
+        }
 
         return Response.status(200).build();
     }
